refactor: extract shared utilities and update Pixel Forge integration
- Create shared utilities for common functionality
- Update Pixel Forge router for consistency with Picture Press
